requirements-dev.txt
# Development dependencies for release automation
google-genai>=0.8.0
build>=0.10.0
twine>=4.0.0